feat(broadcast): combine request and response types
